+++ previous-versions/banking/src/sdk/sdk.ts
@@ -56,8 +56,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.30.1";
-    genVersion = "2.131.1";
+    sdkVersion = "0.31.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 0.31.0 2.139.1 3.0.0 @codat/banking";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);