+++ previous-versions/banking/RELEASES.md
@@ -559,3 +559,13 @@ Based on:
 - [typescript v0.30.1] previous-versions/banking
 ### Releases
 - [NPM v0.30.1] https://www.npmjs.com/package/@codat/banking/v/0.30.1 - previous-versions/banking
+
+## 2023-09-29 00:22:14
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.31.0] previous-versions/banking
+### Releases
+- [NPM v0.31.0] https://www.npmjs.com/package/@codat/banking/v/0.31.0 - previous-versions/banking